src/main/java/ru/job4j/tracker/Tracker.java
package ru.job4j.tracker;

import java.util.ArrayList;
import java.util.List;

public class Tracker {
    private final List<Item> items = new ArrayList<>();
    private int ids = 1;
    //private int size = 0;

    public Item add(Item item) {
        item.setId(ids++);
        items.add(item);
        //size++;
        return item;
    }

    public List<Item> findAll() {
        return List.copyOf(items);
    }

    public Item findById(int id) {
        int index = indexOf(id);
        return  index != -1 ? items.get(index) : null;
    }

    public List<Item> findByName(String key) {
        List<Item> rsl = new ArrayList<>();
        for (Item item : items) {
            if (key.equals(item.getName())) {
                rsl.add(item);
            }
        }
        return rsl;
    }

    private int indexOf(int id) {
        int rsl = -1;

        for (Item item : items) {
            if (item.getId() == id) {
                rsl = items.indexOf(item);
                break;
            }
        }
        return rsl;
    }

    public boolean replace(int id, Item item) {
        int index = indexOf(id);

        if (index != -1) {
            int oldId = items.get(index).getId();
            item.setId(oldId);
            items.set(index, item);
            return true;
        } else {
            return false;
        }
    }

    public boolean delete(int id) {
        int index = indexOf(id);
        if (index != -1) {
           items.remove(index);
           //System.arraycopy(items, index + 1, items, index, size - index - 1);
           //size--;
           return true;
        } else {
            return false;
        }
    }

}